Add borrowing edge iterator.
An iterator over all edges of a graph that contains a reference to the respective `EdgeData`.
This forces the graph traits to carry the lifetime of that iterator around, which is not nice.
HRTBs would force to limit all `EdgeData`s to `'static`, which could be okay.

I realised that I don't need this and implemented a better solution before: I just let the iterator return edge ids, which can then be used to retrieve the data in a separate step, using a direct method.