Implement reading and writing edge-centric from and to bcalm2.
On the way, we tested the code with the test case from the node-centric read/write.
This helped us discover a bug in reading where the neighbor ids of bcalm2 nodes were not mapped correctly to our edge ids. Since we add two edges per bcalm2 node, we need to multiply the neighbor ids from the bcalm2 nodes by 2 to get the correct ids of already added edges.
Moreover, we forgot about various loops that can occur in a node-centric bidirected graph. Those are edges from a binode to itself, with or without changing direction.

Relates to #57