Completely refactor Snakefile and add Bandage.
While adding bandage I encountered problems with mixed file endings, and
as a solution the best idea I had was to just refactor the complete
Snakefile and introduce a better naming scheme. As it happened I then
did the refactoring together with adding Bandage. So no separate
commits for that.

Closes #121