Speed up (and correct) hydrostructure initialisation.
Apparently there was a case that our tests didn't capture, where the walk was initially univocal or R-univocal, and then the vapor_is_path_tracker did not get initialised correctly. It would potentially leave out nodes or edges which are in R⁺ and R⁻ at that point, but not in the vapor. This can happen since the incremental property does not hold for univocal and R-univocal walks.
The vapor is tracked only for R⁺ and R⁻, and walks that do not have splits or do not have joins are handled separately in the queries.

For speeding things up first of all the correction of the initialisation of the vapor_is_path_tracker makes it asymptotically faster, as it does not need to check the whole walk for containing a node or edge anymore.
Also we add methods to initialise the hydrostructure with the wanted fingers right after computation so that we do not need to initialise it twice. Moreover we add a method that initialises the fingers to the leftmost positions and contains more efficient code for this case.

Relates to #137