Remove arcs making the preprocessed graph hamiltonian.
This was a bug before.
If e.g. a safe walk [n1, n2, n3] would be contracted, then n2 was
removed, but the outgoing arcs of n1 and the incoming arcs of n3 were
kept. This could make the resulting walk [n1, n3] in the preprocessed
graph unsafe.

Relates to #179