Load the graph topology using the .3.dot file.
Results in unitigs being much closer to the original. There might still be some too much or missing, but this might be caused by wrong filtering, specifically by computing the base-pair length of a unitig wrongly.

What is still missing are many reads that support edges. It might be required to just add each read supporting an edge to the existing edges and ignore the closed markers in the reads file.

Relates to #185
Relates to #186