Fix all the endless loops in trivial omnitigs in non strongly connected graphs.
The compute-trivial-omnitigs command now features a flag --non-scc to enable the algorithms that can handle non-SC graphs. If it is not used, but the graph is not SC, then the command now fails properly instead of entering an endless loop.

The algorithms that handle not-SC graphs are slightly slower, so if the graph is SCC the flag should not be used. But in practice, it does not seem to make a difference for the graphs we tested so far. They are quadratic in the length of the univocal extension now though, so it might explode in the future. If that happens, the algorithms can still be returned to linear runtime.